# backend/openapi/swagger.yaml
components:
  schemas:
...
    github_com_uptaris_uptaris_backend_internal_request.Credentials:
      properties:
        email:
          example: operator@example.com
          maxLength: 254
          type: string
        password:
          example: UptarisDemo!2026
          maxLength: 128
          minLength: 8
          type: string
      required:
      - email
      - password
      type: object
...
    github_com_uptaris_uptaris_backend_internal_request.PasskeyVerification:
      properties:
        ceremonyToken:
          type: string
        credential:
          type: object
        name:
          type: string
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.AuthResponse:
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.UserResponse'
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.ErrorResponse:
      properties:
        error:
          properties:
            code:
              type: string
            message:
              type: string
          required:
          - code
          - message
          type: object
      required:
      - error
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.PasskeyOptionsResponse:
      properties:
        ceremonyToken:
          type: string
        optionsJSON: {}
      type: object
    github_com_uptaris_uptaris_backend_internal_response.PasskeyResponse:
      properties:
        createdAt:
          type: string
        id:
          type: string
        lastUsedAt:
          type: string
        name:
          type: string
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.SignInResponse:
      properties:
        accessToken:
          type: string
        twoFactorRequired:
          type: boolean
        user:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.UserResponse'
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.UserResponse:
      properties:
        email:
          type: string
        id:
          example: "0"
          type: string
        role:
          type: string
      type: object
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  description: REST API for Uptaris infrastructure monitoring.
  title: Uptaris API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /account/passkeys:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.PasskeyResponse'
                type: array
          description: OK
      security:
      - bearerauth: []
      summary: List account passkeys
      tags:
      - account
  /account/passkeys/{id}:
    delete:
      parameters:
      - description: Passkey ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - bearerauth: []
      summary: Remove account passkey
      tags:
      - account
  /account/passkeys/options:
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.PasskeyOptionsResponse'
          description: OK
      security:
      - bearerauth: []
      summary: Begin passkey enrollment
      tags:
      - account
  /account/passkeys/verify:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_request.PasskeyVerification'
        description: Passkey attestation and one-time ceremony token
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.PasskeyResponse'
          description: Created
      security:
      - bearerauth: []
      summary: Finish passkey enrollment
      tags:
      - account
...
  /auth/login:
    post:
      description: Returns accessToken and refresh cookie, or {"twoFactorRequired":true}
        and HttpOnly challenge cookie when 2FA is enabled. Complete /auth/two-factor/verify
        before using authenticated routes.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_request.Credentials'
        description: Email and password
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.SignInResponse'
          description: OK
          headers:
            Set-Cookie:
              description: HttpOnly refresh cookie
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_json: malformed JSON body'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_credentials: email or password invalid'
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'origin_not_allowed: request origin not allowed'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'unsupported_media_type: non-JSON request body'
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'validation_failed: invalid request field'
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'rate_limited: too many authentication attempts'
          headers:
            Retry-After:
              description: Seconds before retry
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      summary: Log in
      tags:
      - auth
...
  /auth/oauth/{provider}:
    get:
      parameters:
      - description: OAuth provider ID
        in: path
        name: provider
        required: true
        schema:
          type: string
      responses:
        "302":
          description: Redirect to provider authorization
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: Service Unavailable
      summary: Start OAuth sign-in
      tags:
      - auth
  /auth/oauth/{provider}/callback:
    get:
      parameters:
      - description: OAuth provider ID
        in: path
        name: provider
        required: true
        schema:
          type: string
      - description: Provider authorization code
        in: query
        name: code
        required: true
        schema:
          type: string
      - description: One-time state value
        in: query
        name: state
        required: true
        schema:
          type: string
      responses:
        "302":
          description: Redirect to app
      summary: Complete OAuth sign-in or connection
      tags:
      - auth
  /auth/oauth/{provider}/link:
    delete:
      parameters:
      - description: OAuth provider ID
        in: path
        name: provider
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: last_signin_method
      security:
      - bearerauth: []
      summary: Disconnect OAuth provider
      tags:
      - auth
    post:
      parameters:
      - description: OAuth provider ID
        in: path
        name: provider
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: Unauthorized
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: Service Unavailable
      security:
      - bearerauth: []
      summary: Start authenticated OAuth connection
      tags:
      - auth
...
  /auth/passkeys/login/verify:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_request.PasskeyVerification'
        description: Passkey assertion and one-time ceremony token
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.AuthResponse'
          description: OK
      summary: Finish passkey sign-in
      tags:
      - auth
...
servers:
- url: /api/v1